move deletable logic down to the page tree component and refactor page navgiation
