+++ components/common/PageLayout/components/PageNavigation/index.ts
@@ -1,0 +1,3 @@
+import PageNavigation from './PageNavigation';
+
+export default PageNavigation;
